+++ packages/netlify-cms-core/src/components/Editor/EditorInterface.js
@@ -14,8 +14,8 @@ import {
 } from 'netlify-cms-ui-default';
 import { ScrollSync, ScrollSyncPane } from 'react-scroll-sync';
 import EditorControlPane from './EditorControlPane/EditorControlPane';
-import RightControlPane from "./RightControlPane/RightControlPane";
-import EditorNotesPane from "./EditorNotesPane/EditorNotesPane";
+import RightControlPane from './RightControlPane/RightControlPane';
+import EditorNotesPane from './EditorNotesPane/EditorNotesPane';
 import EditorPreviewPane from './EditorPreviewPane/EditorPreviewPane';
 import EditorToolbar from './EditorToolbar';
 import { hasI18n, getI18nInfo, getPreviewEntry } from '../../lib/i18n';
@@ -274,33 +274,34 @@ class EditorInterface extends Component {
     const previewEntry = collectionI18nEnabled
       ? getPreviewEntry(entry, leftPanelLocale, defaultLocale)
       : entry;
-    
-    const tabElements = [{
-      key: "notes",
-      title: "Notes",
-      content: (
-        <EditorNotesPane />
-      )
-    }, {
-      key: "preview",
-      title: "Preview",
-      content: (
-        <EditorPreviewPane
-          collection={collection}
-          entry={previewEntry}
-          fields={fields}
-          fieldsMetaData={fieldsMetaData}
-        />
-      )
-    }]
+
+
+    const tabElements = [
+      {
+        key: 'notes',
+        title: 'Notes',
+        content: <EditorNotesPane />,
+      },
+      {
+        key: 'preview',
+        title: 'Preview',
+        content: (
+          <EditorPreviewPane
+            collection={collection}
+            entry={previewEntry}
+            fields={fields}
+            fieldsMetaData={fieldsMetaData}
+          />
+        ),
+      },
+    ];
 
     const previewPane = (
       <PreviewPaneContainer blockEntry={showEventBlocker}>
         <RightControlPane tabs={tabElements} />
       </PreviewPaneContainer>
-    )
+    );
 
-    
     const editorWithPreview = (
       <ScrollSync enabled={this.state.scrollSyncEnabled}>
         <div>