+++ packages/netlify-cms-core/src/components/Editor/EditorNotesPane/EditorNotesPane.js
@@ -1,9 +1,8 @@
-import React from "react";
-import { connect } from 'react-redux';
+import React from 'react';
 import styled from '@emotion/styled';
 
-import CreateNote from "./CreateNote";
-import NoteList from "./NoteList";
+import CreateNote from './CreateNote';
+import NoteList from './NoteList';
 
 const NotesContainer = styled.div`
   overflow: none;
@@ -18,28 +17,5 @@ export default function EditorNotesPane() {
       <NoteList />
       <CreateNote />
     </NotesContainer>
-  )
+  );
 }
-
-
-// function mapStateToProps(state) {
-//   const isLoadingAsset = selectIsLoadingAsset(state.medias);
-//   return { isLoadingAsset, config: state.config };
-// }
-
-// function mapDispatchToProps(dispatch) {
-//   return {
-//     boundGetAsset: (collection, entry) => boundGetAsset(dispatch, collection, entry),
-//   };
-// }
-
-// function mergeProps(stateProps, dispatchProps, ownProps) {
-//   return {
-//     ...stateProps,
-//     ...dispatchProps,
-//     ...ownProps,
-//     getAsset: dispatchProps.boundGetAsset(ownProps.collection, ownProps.entry),
-//   };
-// }
-
-// export default connect(mapStateToProps, mapDispatchToProps, mergeProps)(EditorNotesPane);